pyproject.toml
[project]
name = "voice-agent"
version = "1.0.0"
description = "Speech-to-text and text-to-speech desktop agent using Gemini and ElevenLabs"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "pyaudio",
    "pynput",
    "pyperclip",
    "python-dotenv",
]

[project.optional-dependencies]
elevenlabs = ["elevenlabs"]
test = [
    "pytest",
    "pytest-asyncio",
    "numpy",
]
dev = [
    "voice-agent[test,elevenlabs]",
    "ruff",
    "mypy",
]

[project.scripts]
voice-agent = "voice_agent.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["voice_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
